[stable10] Remove alternate keys storage during user delete
When user is deleted the alternate key storage location
does have folder with username. So while deleting user
this should also be cleaned up.

+++ apps/encryption/lib/Hooks/UserHooks.php
@@ -198,7 +198,14 @@ class UserHooks implements IHook {
 	public function postDeleteUser($params) {
 
 		if (App::isEnabled('encryption')) {
-			$this->keyManager->deletePublicKey($params['uid']);
+			/**
+			 * Adding a safe condition to make sure the uid is not
+			 * empty or null.
+			 */
+			if (!is_null($params['uid']) && ($params['uid'] !== '')) {
+				$this->keyManager->deletePublicKey($params['uid']);
+				\OC::$server->getEncryptionKeyStorage()->deleteAltUserStorageKeys($params['uid']);
+			}
 		}
 	}
